Big fix: Fixed rotation & translation + other related improvements
-Added output of view plane position parameters
-Changed volume positioning: Volume is now centered around origin
-View plane now rotates around origin and the volume located there

This checkin might need a fixup. Solved multiple problems.

Left and up vectors of view plane did not scale with view plane size, this is now fixed.

Also added transformation that moves volume to the origin, which is a prerequisite for proper rotation & translation behavior.

Finally, fixed translation and rotation by putting view plane in the origin to start, adding a default translation to move it a fixed distance from the origin, and finally swapped the order of translation and rotation to be correct wrt. this particular raycasting scenario.