Load 1D texture to use for transfer function, POC (has bug, see details)
Problem: See edges of transfer function demonstration render. This is supposed to be a smooth gradient from white to black, with a red stripe on the right. The leftmost and rightmost parts are wrong, because the edge of the texture interpolates towards the zero (or incorrect, anyway) values on the edges of the texture.

A new sampler needs to be created, such that these values are clamped to the inside of the texture data.

Had to read up a bit on shader resource views, descriptor heaps and root signatures to figure out how D3D12 loads textures into a common GPU-accessible data structure.